requirements.txt
Django>=6.0.1
djangorestframework>=3.14.0
Pillow>=10.0.0

